read below for info
Changed the ModelJson class so that its up to date and functioning
changed the setfavourite and getfavourite method of ListOfdrinks, so
that the favourite.txt file already exists upon starting the skill.
Regarding the tests for the List of drinks, i tweaked the AssertEquals
tests to fit the new outputs.
Added the new Handlers to the SimpleBarkeeperStreamHandler class.
Added a new slot for the Favourite Slot.
Created a new mockinput Generator in TestUtil to easily mock the Set and
GetFavouriteIntentHandler tests.
Created the Get And SetFavouriteIntentHandlers as well as the respective
TestClasses.